+++ NEWS.md
@@ -4,6 +4,9 @@
 
 * `as_factor()` and `fct_inorder()` accept NA levels (#98).
 
+* `fct_lump()` gains `weights` argument (#70, @wilkox) to weight value
+  frequencies before lumping them together (#68).
+
 # forcats 0.2.0
 
 ## New functions